pyproject.toml
[tool.poetry]
name = "financial-agent"
version = "0.1.0"
description = "A financial analysis AI agent using LangChain and LangGraph to provide insights to users."
authors = ["abcdef@gmail.com"]
readme = "README.md"
packages = [{include = "financial_agent"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain-community = "0.0.13"
langchain-core = "0.1.10"
langchain-openai = "0.0.2.post1"
yfinance = "0.2.35" 
scikit-learn = "1.3.2"
scipy = "1.11.4"
ta = "0.11.0"
numpy = "1.24.3"
pandas = "2.0.3"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
